feat: Add EncryptionService and fix errors in existing classes
